+++ requirements.txt
@@ -1,0 +1,71 @@
+absl-py==2.0.0
+aiodns==3.1.1
+aiohttp==3.8.6
+aiosignal==1.3.1
+astunparse==1.6.3
+async-timeout==4.0.3
+attrs==23.1.0
+binance==0.3
+binance.py==1.9.0
+cachetools==5.3.1
+ccxt==4.1.22
+certifi==2023.7.22
+cffi==1.16.0
+charset-normalizer==3.3.1
+cryptography==41.0.4
+flatbuffers==23.5.26
+frozenlist==1.4.0
+gast==0.5.4
+gevent==23.9.1
+google-auth==2.23.3
+google-auth-oauthlib==1.0.0
+google-pasta==0.2.0
+greenlet==3.0.0
+grpcio==1.59.0
+h5py==3.10.0
+idna==3.4
+joblib==1.3.2
+keras==2.14.0
+libclang==16.0.6
+Markdown==3.5
+MarkupSafe==2.1.3
+ml-dtypes==0.2.0
+multidict==6.0.4
+numpy==1.26.1
+oauthlib==3.2.2
+opt-einsum==3.3.0
+packaging==23.2
+pandas==2.1.1
+protobuf==4.24.4
+pyasn1==0.5.0
+pyasn1-modules==0.3.0
+pycares==4.4.0
+pycparser==2.21
+pyTelegramBotAPI==4.14.0
+python-dateutil==2.8.2
+pytz==2023.3.post1
+requests==2.31.0
+requests-oauthlib==1.3.1
+rsa==4.9
+scikit-learn==1.3.1
+scipy==1.11.3
+six==1.16.0
+ta==0.10.2
+telebot==0.0.5
+tensorboard==2.14.1
+tensorboard-data-server==0.7.1
+tensorflow==2.14.0
+tensorflow-estimator==2.14.0
+tensorflow-intel==2.14.0
+tensorflow-io-gcs-filesystem==0.31.0
+termcolor==2.3.0
+threadpoolctl==3.2.0
+typing_extensions==4.8.0
+tzdata==2023.3
+urllib3==2.0.7
+websocket==0.2.1
+Werkzeug==3.0.0
+wrapt==1.14.1
+yarl==1.9.2
+zope.event==5.0
+zope.interface==6.1
